requirements.txt
asyncio
httpx[http2]
python-dotenv